feat: change admin dashboard's access
