--- client/src/modules/AdminDashboard/index.tsx
@@ -1,1 +1,0 @@
-export { default as AdminDashboard } from './pages/AdminDashboard';
