+++ client/src/modules/CourseStatistics/components/EpamMentorsStatsCard/index.tsx
