+++ client/src/modules/CourseStatistics/components/MentorsCountriesCard/MentorsCountriesCard.tsx
@@ -1,14 +1,16 @@
 import { Card } from 'antd';
 import { CountriesStatsDto } from 'api';
 import dynamic from 'next/dynamic';
-import { Colors } from '../data';
+import { Colors } from '../../data';
 
 type Props = {
   countriesStats: CountriesStatsDto;
   activeCount: number;
 };
 
-const CountriesChart = dynamic(() => import('../CountriesChart/CountriesChart'), { ssr: false });
+const CountriesChart = dynamic(() => import('../CountriesChart/CountriesChart'), {
+  ssr: false,
+});
 
 export const MentorsCountriesCard = ({ countriesStats, activeCount }: Props) => {
   const { countries } = countriesStats;
