+++ client/src/modules/CourseStatistics/components/StudentsStatsCard/index.tsx
