+++ client/src/modules/CourseStatistics/components/StudentsWithCertificateCard/StudentsWithCertificateCard.tsx
@@ -1,7 +1,7 @@
 import { Card, Typography } from 'antd';
 import { CourseStatsDto } from 'api';
 import dynamic from 'next/dynamic';
-import { Colors } from '../data';
+import { Colors } from '../../data';
 
 type Props = {
   studentsStats: CourseStatsDto;
@@ -9,17 +9,19 @@ type Props = {
 
 const { Text } = Typography;
 
-const LiquidChart = dynamic(() => import('../LiquidChart/LiquidChart'), { ssr: false });
+const LiquidChart = dynamic(() => import('../LiquidChart/LiquidChart'), {
+  ssr: false,
+});
 
-export const StudentsEligibleForCertificationCard = ({ studentsStats }: Props) => {
+export const StudentsWithCertificateCard = ({ studentsStats }: Props) => {
   return (
-    <Card title="Eligible for Certification">
+    <Card title="Students With Certificate">
       <Text strong>
-        Eligible for Certification: {studentsStats.eligibleForCertificationCount} / {studentsStats.activeStudentsCount}
+        Students With Certificate: {studentsStats.certifiedStudentsCount} / {studentsStats.activeStudentsCount}
       </Text>
       <div style={{ height: 180, width: '100%' }}>
         <LiquidChart
-          count={studentsStats.eligibleForCertificationCount}
+          count={studentsStats.certifiedStudentsCount}
           total={studentsStats.activeStudentsCount}
           color={Colors.Lime}
         />
