+++ client/src/modules/CourseStatistics/data.ts
