+++ client/src/modules/CourseStatistics/hooks/index.ts
