+++ client/src/modules/CourseStatistics/index.tsx
@@ -1,0 +1,1 @@
+export { default as CourseStatistics } from './pages/CourseStatistics';
